build: integrate unit testing
