+++ .gitignore
@@ -1,7 +1,5 @@
-webserv
-*.o
-*.d
+build/
 tester
-cgi_tester
 ubuntu_tester
+cgi_tester
 ubuntu_cgi_tester
