+++ Makefile
@@ -1,13 +1,13 @@
-# Executable name
+#################################################################################
+# Project setup
+
 NAME      = webserv
 
-# Directories
 SRC_DIR   = src
 BUILD_DIR = build
 OBJ_DIR   = $(BUILD_DIR)/obj
 BIN_DIR   = $(BUILD_DIR)/bin
 
-# Main target
 TARGET    = $(BIN_DIR)/$(NAME)
 
 # Source tree (keep in alphabetical order)
@@ -17,6 +17,8 @@ SRCS      = $(addprefix $(SRC_DIR)/, \
             core/client.hpp \
             core/server.cpp \
             core/server.hpp \
+            core/signals.cpp \
+            core/signals.hpp \
             http/http_request.hpp \
             http/http_response.cpp \
             http/http_response.hpp \
@@ -40,6 +42,12 @@ CXX       = clang++
 CXXFLAGS  = -Wall -Wextra -Werror -std=c++98 -g3
 CPPFLAGS  = -I$(SRC_DIR) -MMD -MP
 
+# For utest.h
+CPPFLAGS  += -Ithird_party/utest
+
+#################################################################################
+# Default build
+
 # Default target
 PHONY += all run val
 all: $(TARGET)
@@ -58,7 +66,7 @@ $(OBJ_DIR)/%.o: $(SRC_DIR)/%.cpp
 	@mkdir -p $(dir $@)
 	$(CXX) $(CXXFLAGS) $(CPPFLAGS) -c -o $@ $<
 
-# Regular cleaning targets
+# Default cleaning targets
 PHONY += clean fclean re
 clean:
 	rm -rf $(OBJ_DIR)
@@ -68,11 +76,54 @@ fclean: clean
 
 re: fclean all
 
+#################################################################################
+# Unit tests
+
+TEST_NAME    = run_tests
+
+TEST_SRC_DIR = tests
+TEST_OBJ_DIR = $(BUILD_DIR)/obj_tests
+
+TEST_TARGET  = $(BIN_DIR)/$(TEST_NAME)
+
+# All tests sources (keep in alphabetical order)
+TEST_SRCS    = $(addprefix $(TEST_SRC_DIR)/, \
+               http_response_unittest.cpp \
+               main.cpp \
+)
+
+TEST_CPPS    = $(filter %.cpp,$(TEST_SRCS))
+
+TEST_OBJS    = $(patsubst $(TEST_SRC_DIR)/%,$(TEST_OBJ_DIR)/%,$(TEST_CPPS:.cpp=.o))
+TEST_DEPS    = $(TEST_OBJS:.o=.d)
+
+DEPS         += $(TEST_DEPS)
+
+# Test runner target
+PHONY += test
+test: $(TEST_TARGET)
+	./$(TEST_TARGET)
+
+$(TEST_TARGET): $(TEST_OBJS) $(filter-out $(OBJ_DIR)/main.o,$(OBJS))
+	@mkdir -p $(dir $@)
+	$(CXX) $(CXXFLAGS) $(CPPFLAGS) $^ -o $@
+
+$(TEST_OBJ_DIR)/%.o: $(TEST_SRC_DIR)/%.cpp
+	@mkdir -p $(dir $@)
+	$(CXX) $(CXXFLAGS) $(CPPFLAGS) -c -o $@ $<
+
+# Tests cleaning target
+PHONY += test-clean
+test-clean:
+	rm -rf $(TEST_OBJ_DIR) $(TEST_TARGET)
+
+#################################################################################
 # Formatting and linting section
 
 # If compiledb is not installed do `pipx install compiledb`.
 # compiledb is used to generate the compilation database (compile_commands.json)
 # used by clang-tidy.
+
 PHONY += db format lint check
 db:
 	compiledb -n $(MAKE)
@@ -83,7 +134,7 @@ format:
 lint: db
 	clang-tidy -p=. --header-filter=.* --warnings-as-errors=* $(CPPS)
 
-check: format lint
+check: format lint test
 
 # Careful, these targets will overwrite files.
 # Make sure to use `make check` before committing irreversible changes.
@@ -95,6 +146,8 @@ lint-fix: db
 	clang-tidy -p=. --header-filter=.* -fix $(CPPS)
 
 fix: format-fix lint-fix
+
+
 
 # Include dependencies
 -include $(DEPS)
