+++ src/core/server.hpp
@@ -10,11 +10,7 @@
 #include <sys/epoll.h>
 #include <sys/socket.h>
 
-#include <csignal>
-
 #define WEBSERV_MAX_EVENTS 64
-
-extern volatile sig_atomic_t g_sigint_received;
 
 class Server {
 public:
