+++ tests/main.cpp
@@ -1,0 +1,6 @@
+#include "utest/utest.h"
+
+// Don't need to do much here
+// https://github.com/sheredom/utest.h?tab=readme-ov-file#utest_main
+
+UTEST_MAIN();
